linux: Remove use of the lsl instruction
The lsl instruction is no longer supported by binutils.  I believe that
this was lost during the big binutils rewrite.  We could try to add this
back, however, the upstream linux kernel code has been modified to
remove use of lsl.   That and lsl is not mentioned in any ARC manuals
would make selling lsl a bit tricky (unless we added it as an nps only
instruction alias).

In this commit I am removing use of lsl from the linux kernel, which is
inline with the current kernel code.