linux: Increase size of CONFIG_BLK_DEV_RAM_SIZE
Was hitting a case where unpacking the rootfs image would not work
correctly, but if I made a single file 1 byte smaller then the image
would unpack correctly.

Searching the web leads me to believe that what is happening is that
the space set aside for the ramfs is being exceeded, leading to a
corrupt file system image.

It is not entirely clear if it is safe to increase this config value
for the system being simulated, but this seems to leave me in a more
stable position than when I started.