uclibc: Support more relocations during bootstrap phase
During the boostrap phase of the runtime linker we currently only
support R_ARC_RELATIVE relocations.  The problem is that an R_ARC_NONE
relocation is generated by default, and if uClibc is configured with
'DODEBUG=y' then we also generate an R_ARC_JMP_SLOT relocation too.

This patch adds support for these relocations during the bootstrap phase.