uclibc: Avoid using @gotpc with local symbols
This patch changes uclibc so that we avoid using @gotpc with a local
symbol.  The @gotpc style symbol references generate the difference
between the current pc and the entry for the given symbol in the GOT.
The problem is that local symbols don't have an entry in the GOT, and
so, what actually happens is that binutils appears to get a little
confused.

The linker allocates a space in the GOT for the local symbol reference,
but does not actually arrange to have this space filled with the address
of the local symbol.  Instead, we are left with an R_ARC_NONE
relocation, and an entry in the GOT that will be undefined (though
probably zero in practice).

We only used the @gotpc style reference with a local symbol on one code
path, an alternative code path used a different, global symbol.  This
new patch unifies on the single global symbol code path.  As a result we
should no longer be generating R_ARC_NONE relocation in the GOT for the
dynamic linker.